+++ RELEASE_NOTES.md
@@ -8,6 +8,10 @@
 
 * удален NotificationCenter - используйте `PushHandler`
 
+#### util-ktx
+* ANDDEP-319 Свойства `isAtLeast...`  класса [`SdkUtils`](util-ktx/src/main/java/ru/surfstudio/android/utilktx/util/SdkUtils.kt)
+помечены как @Deprecated, вместо них следует использовать методы `isAtLeast...()`, а так же `runOn...()`.
+
 ## 0.3.0
 
 #### core-ui